moving around some helpers
